AI/requirements.txt
google-genai>=0.2.0
google-auth>=2.23.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
boto3>=1.34.0
requests>=2.31.0
